refactor: move the event log into the analytics app
Event was the last big platform model sitting in the edition — it records
what users do across the whole product, and gyrinx.analytics existed
already with an empty models.py and a comment saying it needed no
database. It does now.

The model turned out to be clean: its only dependencies were AppBase
(moved to the platform in the previous commit) and Django's ContentType.
Its `object` GenericForeignKey does point at edition rows, but loosely, by
content-type id — that is a reference, not a schema dependency, and it
keeps working untouched.

Same technique as the other model moves: db_table pinned to core_event,
migrations wrapped in SeparateDatabaseAndState with no database
operations, ContentType row carried across in place.

This also unblocks csrf_failure, which needed log_event and was the last
settings entry still pointing into n23 for a non-model reason.

Verified against a populated database: 1,532 event rows before and after,
all ten indexes still on core_event, no analytics_* table created. 4166
tests pass, makemigrations --check clean.

Refs #2093
