+++ bandit/bandit-baseline.txt
@@ -1,4 +1,4 @@
-Run started:2026-08-04 22:24:48.501976+00:00
+Run started:2026-08-04 22:28:18.540114+00:00
 
 Test results:
 >> Issue: [B703:django_mark_safe] Potential XSS on mark_safe function.
@@ -340,7 +340,7 @@ Test results:
 --------------------------------------------------
 
 Code scanned:
-	Total lines of code: 71206
+	Total lines of code: 71207
 	Total lines skipped (#nosec): 4
 	Total potential issues skipped due to specifically being disabled (e.g., #nosec BXXX): 56
