+++ n23/core/models/__init__.py
@@ -3,7 +3,6 @@ from .backfill import *  # noqa: F403
 from .battle import *  # noqa: F403
 from .campaign import *  # noqa: F403
 from .crew import *  # noqa: F403
-from .events import *  # noqa: F403
 from .facts import *  # noqa: F403
 from .invitation import *  # noqa: F403
 from .list import *  # noqa: F403
