+++ n23/core/tests/test_event_list_update_signal.py
@@ -5,7 +5,7 @@ from datetime import timedelta
 import pytest
 from django.utils import timezone
 
-from n23.core.models.events import Event, EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import Event, EventNoun, EventVerb, log_event
 from n23.core.models.list import List
