+++ n23/core/tests/test_impersonation.py
@@ -15,7 +15,7 @@ from gyrinx.impersonation import (
     can_impersonate_target,
 )
 from n23.core.middleware import ImpersonationMiddleware
-from n23.core.models import Event
+from gyrinx.analytics.models import Event
 from gyrinx.site.models import ImpersonationLog
