+++ n23/core/views/campaign/arbitrators.py
@@ -11,7 +11,7 @@ from django.views.decorators.http import require_POST
 from gyrinx import messages
 from n23.core.forms.campaign import AddArbitratorForm
 from n23.core.models.campaign import CampaignAction
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.views.campaign.common import get_campaign_admin_or_404
