+++ n23/core/views/campaign/assets.py
@@ -22,7 +22,7 @@ from n23.core.models.campaign import (
     CampaignAssetType,
     CampaignSubAsset,
 )
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.utils import get_return_url, safe_redirect
 from gyrinx.tracker import track
 from n23.core.views.campaign.common import get_campaign_admin_or_404
