+++ n23/core/views/campaign/attributes.py
@@ -21,7 +21,7 @@ from n23.core.models.campaign import (
     CampaignAttributeType,
     CampaignAttributeValue,
 )
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.utils import get_return_url, safe_redirect
 from n23.core.views.campaign.common import get_campaign_admin_or_404
