+++ n23/core/views/campaign/copy.py
@@ -13,7 +13,7 @@ from n23.core.handlers.campaign_copy import (
     copy_campaign_content,
 )
 from n23.core.models.campaign import Campaign
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from gyrinx.tracker import track
 from n23.core.views.campaign.common import get_campaign_admin_or_404
