+++ n23/core/views/fighter/advancements.py
@@ -29,7 +29,7 @@ from n23.core.handlers.fighter import (
     handle_fighter_advancement_deletion,
 )
 from n23.core.models.campaign import CampaignAction
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.models.list import (
     List,
     ListFighter,
