+++ n23/core/views/fighter/crud.py
@@ -24,7 +24,7 @@ from n23.core.handlers.fighter import (
     handle_fighter_kill,
     handle_fighter_resurrect,
 )
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.models.list import List, ListFighter
 from n23.core.views.list.common import get_clean_list_or_404
