+++ n23/core/views/fighter/narrative.py
@@ -8,7 +8,7 @@ from n23.core.forms.list import (
     EditListFighterNarrativeForm,
     EditListFighterNotesForm,
 )
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.models.list import List, ListFighter
 from n23.core.utils import get_return_url, safe_redirect
 from n23.core.views.list.common import get_clean_list_or_404
