+++ n23/core/views/fighter/rules.py
@@ -11,7 +11,7 @@ from django.urls import reverse
 from gyrinx import messages
 from n23.core.utils import search_queryset
 from n23.content.models import ContentRule
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.models.list import List, ListFighter
 from n23.core.views.list.common import get_clean_list_or_404
 from gyrinx.models import QuerySetOf, is_valid_uuid
