+++ n23/core/views/home.py
@@ -9,7 +9,7 @@ from n23.core.utils import search_queryset
 from n23.content.models import ContentHouse
 from n23.core.models.battle import Battle
 from n23.core.models.campaign import Campaign
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.models.list import List, ListFighter
 from n23.core.models.pack import CustomContentPack
 from gyrinx.tracing import span, traced
