+++ n23/core/views/list/skill_trees.py
@@ -9,7 +9,7 @@ from django.urls import reverse
 
 from gyrinx import messages
 from n23.core.forms.skill_tree import ListSkillTreeForm
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.models.list import List
 from n23.core.utils import get_return_url, safe_redirect
 from n23.core.views.list.common import get_clean_list_or_404
