+++ n23/core/views/vehicle.py
@@ -23,7 +23,7 @@ from n23.core.forms.vehicle import (
     VehicleSelectionForm,
 )
 from n23.core.handlers.fighter import handle_vehicle_purchase
-from n23.core.models.events import EventNoun, EventVerb, log_event
+from gyrinx.analytics.models import EventNoun, EventVerb, log_event
 from n23.core.models.list import List
